Change stylesheets links
Some console 404 errors appeared when clicking on about or comunities sections related to the css files. Also deleted the link with the page.slug variable because it wasn't being used, I thought it could be useful for other sections but not by now because I pretend to add the styles on the liquid file, it depends on their length later on.