Change about.liquid file
This commit applies changes made to other files (services and usecases) in order to reduce repetead code. Specifically: 1. Moved cta class to common.css so it can be used everywhere. 2. Added properties and structure that imitates properties and structure from the previous mencioned files. 3. Tries to unify font weights, font-sizes, margins and gaps. 4. Adds a new class for bigger headers. 5. Moves some data to the data.yml file. 6. Changes the way the last section of Sobre Virto is setted. PR wont be opened until the review from the communication team is given and the refactor reaches the main page